+++ Sources/libcsuper/share.c
@@ -2,7 +2,7 @@
  * \file    share.c
  * \brief   Essential function of libcsuper
  * \author  Remi BERTHO
- * \date    04/07/14
+ * \date    05/07/14
  * \version 4.0.1
  */
 
@@ -147,65 +147,6 @@ int compareFloatDescending(void const *a, void const *b)
    return ret;
 }
 
-/*!
- * \fn FILE *openFile(char file_name[], char mode[])
- *  Open a file with his name and with a specific mode.
- * \param[in] file_name[] the filename
- * \param[in] mode[] the mode
- * \return a pointer to the open file, NULL if there was a problem
- */
-FILE *openFile(char file_name[], char mode[])
-{
-    FILE *ptr_file;
-    libcsuper_initialize();
-
-    if ((ptr_file=fopen(file_name,mode)) == (FILE *) NULL )
-    {
-        printf(_("\nError while opening %s\n"),file_name);
-        perror("");
-    }
-    return ptr_file;
-}
-
-/*!
- * \fn int closeFile(FILE *ptr_file)
- *  Close the file
- * \param[in] *ptr_file the file
- * \return 0 if everything is OK, 1 otherwise
- */
-int closeFile(FILE *ptr_file)
-{
-    int i;
-
-    libcsuper_initialize();
-
-    i=fclose(ptr_file);
-    if (i)
-    {
-        printf(_("Error while closing the file"));
-        perror("");
-    }
-
-    return i;
-}
-
-/*!
- * \fn int readFileSize(FILE *ptr_file)
- *  Read the size of the file
- * \param[in] *ptr_file the file
- * \return the size of the file
- */
-int readFileSize(FILE *ptr_file)
-{
-    int size;
-    int rank=ftell(ptr_file);
-
-    fseek(ptr_file,0,SEEK_END);
-    size=ftell(ptr_file);
-    fseek(ptr_file,rank,SEEK_SET);
-
-    return size;
-}
 
 /*!
  * \fn void *myAlloc(int size_alloue)
@@ -246,73 +187,7 @@ void myRealloc(void **ptr,int size_alloue)
     }
 }
 
-/*!
- * \fn void addFileCsuExtension(char *file_name)
- *  Add the csu file extension
- * \param[in] file_name the filename
- */
-void addFileCsuExtension(char *file_name)
-{
-    char file_extension[4]=FILE_EXTENSION;
-    char ext[4]="abc";
-    int i;
 
-    /*Read the extension of the file*/
-    for (i=strlen(file_name)-3 ; i<strlen(file_name) ; i++)
-            ext[-strlen(file_name)+i+3]=file_name[i];
-
-    /*Add the csu extension if it is not there*/
-    if (strcmp(file_extension,ext)!=0)
-        sprintf(file_name,"%s.%s",file_name,FILE_EXTENSION);
-}
-
-/*!
- * \fn int deleteFile(char *file_name)
- *  Delete a file
- * \param[in] *file_name the filename
- * \return MY_TRUE if everything is OK, MY_FALSE otherwise
- */
-int deleteFile(char *file_name)
-{
-    libcsuper_initialize();
-
-    if(remove(file_name))
-    {
-        printf(_("\nThe file %s cannot be deleted.\n"),file_name);
-        perror("");
-        return MY_FALSE;
-    }
-
-    else
-    {
-        printf(_("\nThe file %s was well deleted.\n"),file_name);
-        return MY_TRUE;
-    }
-}
-
-/*!
- * \fn int renameFile(char *old_name, char *new_name)
- *  Rename a file.
- * \param[in] *old_name the old name of the file
- * \param[in] *new_name the new name of the file
- * \return MY_TRUE if everything is OK, MY_FALSE otherwise
- */
-int renameFile(char *old_name, char *new_name)
-{
-    libcsuper_initialize();
-
-    if(rename(old_name,new_name))
-    {
-        printf(_("\nThe file %s cannot be renamed.\n"),old_name);
-        return MY_FALSE;
-    }
-
-    else
-    {
-        printf(_("\nThe file %s was well renamed in %s.\n"),old_name,new_name);
-        return MY_TRUE;
-    }
-}
 
 /*!
  * \fn char *integerToYesNo(int i, char *yes, char *no)
@@ -330,57 +205,3 @@ char *integerToYesNo(int i, char *yes, char *no)
         return no;
 }
 
-/*!
- * \fn int getFolderFromFilename(char *file_name_to_folder)
- *  Transform a filename into his folder
- * \param[in] file_name_to_folder the filename
- * \return MY_TRUE if everything is OK, MY_FALSE otherwise
- */
-int getFolderFromFilename(char *file_name_to_folder)
-{
-    int i;
-    int ok = MY_FALSE;
-    for (i=strlen(file_name_to_folder) ; i>=0 ; i--)
-    {
-        #ifdef _WIN32
-        if (file_name_to_folder[i] == '\\' || file_name_to_folder[i] == '/')
-        #else
-        if (file_name_to_folder[i] == '/')
-        #endif // _WIN32
-        {
-            ok = MY_TRUE;
-            file_name_to_folder[i] = '\0';
-            break;
-        }
-    }
-
-    return ok;
-}
-
-/*!
- * \fn int getSimpleFilenameFromFullFilename(char *full_filename,char *simple_filename)
- *  Transform a full filename into his simple filename (without the folder)
- * \param[in] full_filename the full filename
- * \param[in] simple_filename the full filename
- * \return MY_TRUE if everything is OK, MY_FALSE otherwise
- */
-int getSimpleFilenameFromFullFilename(char *full_filename,char *simple_filename)
-{
-    int i;
-    int ok = MY_FALSE;
-    for (i=strlen(full_filename) ; i>=0 ; i--)
-    {
-        #ifdef _WIN32
-        if (full_filename[i] == '/' || full_filename[i] == '\\')
-        #else
-        if (full_filename[i] == '/')
-        #endif // _WIN32
-        {
-            ok = MY_TRUE;
-            strcpy(simple_filename,full_filename+i+1);
-            break;
-        }
-    }
-
-    return ok;
-}